Create  login.php
The login.php file contains PHP code responsible for handling the authentication process. It performs the following tasks:
Validates User Credentials: Upon receiving the submitted login form with the username and password, the PHP code validates the provided credentials against a predefined set of authorised users. This step ensures that only authorised users can access the network monitoring tool.
Establishes Server Connection and Redirects: If the provided credentials are correct, the PHP code establishes a connection to the server and grants access to the authenticated user. Subsequently, the user is redirected to the form page, packets.html, where they can specify their preferences for viewing packet data.
Executing "packets.sh" Script: One of the essential functionalities of login.php is to execute the packets.sh shell script every time a user refreshes the page or logs in. This ensures that the captured packet data remains up-to-date and relevant for the user's analysis. The PHP code uses appropriate server-side commands or functions to trigger the execution of the packets.sh script, which captures and processes network packets.

By incorporating this functionality, the network monitoring tool ensures that users have access to real-time and relevant packet data, empowering them to monitor and analyse the network effectively. The seamless integration of PHP with the shell script facilitates a dynamic and responsive user experience.