Create DataBase.txt
To begin, we need to create a MySQL user with appropriate privileges to access the database. This user will be used by the PHP script to connect to the MySQL server and perform necessary operations. The user creation can be done using the following SQL command:


Next, we need to define the table structure in the MySQL database to store the user credentials. The table can be created using the following SQL command:


Once the table is created, you can insert sample entries to have data for testing the login functionality. The entries can be inserted using the following SQL command.